added questionpro tables mappings

+++ src/main/java/com/boot/questionpro/repo/AnswerRepo.java
@@ -1,9 +1,8 @@
 package com.boot.questionpro.repo;
 
+import com.boot.questionpro.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.boot.questionpro.entity.Answer;
-
-public interface AnswerRepo extends JpaRepository<Answer, Long>{
+public interface AnswerRepo extends JpaRepository<Answer, Integer> {
 
 }
