+++ src/main/java/com/boot/questionpro/repo/CXTransactionRepo.java
@@ -1,9 +1,8 @@
 package com.boot.questionpro.repo;
 
+import com.boot.questionpro.entity.CXTransaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.boot.questionpro.entity.CXTransaction;
-
-public interface CXTransactionRepo extends JpaRepository<CXTransaction, Long>{
+public interface CXTransactionRepo extends JpaRepository<CXTransaction, Integer> {
 
 }
