+++ src/main/java/com/boot/questionpro/repo/PanelMemberRepo.java
@@ -1,9 +1,8 @@
 package com.boot.questionpro.repo;
 
+import com.boot.questionpro.entity.PanelMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.boot.questionpro.entity.PanelMember;
-
-public interface PanelMemberRepo extends JpaRepository<PanelMember, Long>{
+public interface PanelMemberRepo extends JpaRepository<PanelMember, Integer> {
 
 }
