+++ src/main/java/com/boot/questionpro/repo/TextResultRepo.java
@@ -1,9 +1,8 @@
 package com.boot.questionpro.repo;
 
+import com.boot.questionpro.entity.TextResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.boot.questionpro.entity.TextResult;
-
-public interface TextResultRepo extends JpaRepository<TextResult, Long>{
+public interface TextResultRepo extends JpaRepository<TextResult, Integer> {
 
 }
